+++ packages/audioplayers_linux/linux/audio_player.cc
@@ -72,6 +72,8 @@ void AudioPlayer::SetSourceUrl(std::string url) {
                 }
             }
         }
+    } else {
+        this->OnPrepared(true);
     }
 }